move colors to tokens
